requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
rembg==2.0.56
Pillow==10.4.0
pillow-avif-plugin==1.4.6
python-multipart==0.0.9
